Accumulated work on fsc.
adds the following new options.

  -ipv4 Use IPv4 rather than IPv6 for the server socket absolute-cp Make
  -classpath elements absolute paths before sending to server max-idle
  -Set idle timeout in minutes for fsc (use 0 for no timeout)

My question marks are what are the right defaults for the first two.
Former behavior is to absolutize the classpath always and never prefer
IPv4 sockets. I changed the default to not absolutize the classpath,
with the option if you need it; I left the system default in place for
the socket creation, but I have a feeling we should default to IPv4. My
only hesitation is that the only way to request an IPv4 socket from java
involves mutating a global system property. (Robustness FTW.) So for
now, you have to give -ipv4.

Closes #3626, #3785, #3788, #3789. Review by community.