Trying to get fsc doing the right thing with re...
Trying to get fsc doing the right thing with respect to absolute and
relative paths. My knowledge of the problem had heretofore been second
hand, and my understanding of it incomplete. The real problem I have
determined is that there are a bunch of different things which go wrong
if relative paths start being resolved from a different base, each of
which needs custom handling.

  classpath-style options, e.g. fsc -cp ../foo.jar
  path-style options, e.g. fsc -d ../mydir
  file arguments, e.g. fsc ../foo.scala

So it was more work than I had realized, or I probably wouldn't have
even touched it. But now it seems to be working as one would want. I
also poured some readability onto the fsc help output.

Closes #4395, no review, but community input would be great.