There is still all kinds of code in the pattern...
There is still all kinds of code in the pattern matcher which I've
been dragging forward (for years now) unclear on the need for it but
having been burned too many times by its undocumented ways. Returned to
the fray once again and encountered shocking success levels with the
abatement stick. Dumped all kinds of unnecessary code (or so we can
hope) and seem close to fixing some fundamental issues. But we've heard
that before. No review.