Improved the error message for another of the m...
Improved the error message for another of the most common situations
I hear about in newbieland. It could be taken further. If compilation
fails due to an unimplemented abstract method, and there is a concrete
method of the same name and arity, it will do a pairwise analysis of the
parameters and attempt to further explain where you went off the beam if
it feels it can do so sensibly. Such as in the test case:

  % scalac S.scala
  S.scala:1: error: class S needs to be abstract, since method g in class J of type (y: Int,z: java.util.List)Int is not defined
  (Note that java.util.List does not match java.util.List[String]. To implement a raw type, use java.util.List[_])
  class S extends J {
        ^
  one error found

No review.