Moved scala.reflect.Chars to scala.reflect.inte...
Moved scala.reflect.Chars to scala.reflect.internal.Chars, un-splitting
the scala.reflect package. This is necessary for running in an OSGi
container, such as Eclipse. A package may not be contributed by two
different bundles (in this case the library and the compiler). It may
be moved back when 'scala.reflect' exists only in one place. review by
extempore.