Overcoming limitations of the repl, multiple-tr...
Overcoming limitations of the repl, multiple-tree lines and assignments.

The best part of nicely documented limitations is that you can stare at
the same comment for years and then the day comes when you give it the
heave-ho.

Exhibit A (regarding a single bare expression): "This is necessary due
to it being hard to modify code at a textual level, and it being hard
to submit an AST to the compiler." Yes, it IS hard to modify code at
a textual level (see enclosed patch) but we won't let that stop us
anymore. Closes #4578.

Exhibit B (regarding the left hand side of assignments having to be
Ident): "An unfortunate limitation." Unfortunate no longer! No review.