Modified return type inference not to allow T* ...
Modified return type inference not to allow T* to leak from varargs
methods. Since I don't know what is supposed to be done about eta
expansion of these methods, I left the behavior as it was (except the
return type) and a boolean val in Types to change it.

  def id[T](xs: T*) = xs
  etaExpandKeepsStar = true   // (id[Int] _) is Int* => Seq[Int]
  etaExpandKeepsStar = false  // (id[Int] _) is Seq[Int] => Seq[Int]

References #4176, leaving open pending resolution of eta expansion.
Review by odersky.