Created simple infrastructure for creating muta...
Created simple infrastructure for creating mutable sets and maps which
are automatically cleared after each compilation run. Since I am not too
familiar with the mechanics of the presentation compiler I'm not sure
this addresses the problem, or that it doesn't clear something which
shouldn't be cleared. Also, this is only a sampling of possible mutable
sets and maps: let me know if it does the job and I can expand it.

Review by dragos.

Conflicts:

	src/compiler/scala/reflect/internal/SymbolTable.scala
	src/compiler/scala/tools/nsc/backend/icode/ICodes.scala

Merge note: Oops, perRunCaches wasn't in 2.9.1.  This is way
more involved than I'd intended, I jumped to the conclusion it
was already in there based on a note in 0c2f493804db6b which
said it should go into 2.9.2.  I think it's OK though.