Fixed a big bug in type constructor unification...
Fixed a big bug in type constructor unification caused by considering
only the parents rather than all the base types. This fix is a testament
to the power of haranguing people in bars when you are deeply offended
by a bug, like someone was by this one:

  def f[CC[X] <: Traversable[X]](x: CC[Int]) = ()
  f(1 to 5)   // did not compile! Fear not, it does now

Review by moors.