Workaround for scalacheck calling System.exit d...
Workaround for scalacheck calling System.exit during testing. Not sure
exactly what the sequence of events is here: I found the exit problem
months ago and committed code to scalacheck to make that configurable,
and then didn't see it again until recently, but I don't think it was
ever addressed. This is just a bandaid. No review.