Fixed a subset of extant signature issues.
When generating signatures, one must be careful finding the name of a
class, because sometimes things with a name like "Foo" need to appear in
signatures as "Foo$". I could really use some help establishing tests
here. Here's the diff of RedBlack's javap:

< const #81 = Asciz	()Lscala/collection/immutable/RedBlack<TA;>.Empty;;
---
> const #81 = Asciz	()Lscala/collection/immutable/RedBlack<TA;>.Empty$;;

RedBlack and RedBlack$Empty$ are the only classfiles in the library
outside of swing which are affected, which nicely explains their
frequent appearance in past tickets.

Review by grek.