Always escape control characters when printing ...
Always escape control characters when printing trees.

Escaping of control characters should be performed for both Char and
String constants. This is useful when printing trees of a program that
contains control character as constants. Before that fix, one could get
really broken output that was hard to read.

Review by extempore.