When I tried to remove the old for comprehensio...
When I tried to remove the old for comprehension syntax in r24958, I
managed to instead only remove the deprecation warnings. No wonder it
went so smoothly. Sorry to get your hopes up only to execute you, old
syntax. Now the for comprehensions have to get it right:

  for (x <- 1 to 5 ; y = x) yield x+y           // nope
  for (val x <- 1 to 5 ; y = x) yield x+y       // nope
  for (val x <- 1 to 5 ; val y = x) yield x+y   // nope
  for (x <- 1 to 5 ; val y = x) yield x+y       // that's the one

No review.