Looks like there was a little baby in the parse...
Looks like there was a little baby in the parser bathwater, or at least
some bathwater still in the general shape of a baby. Apparently both
old forms weren't deprecated. The spec does not suggest that the 'val'
in for comprehension assignment is optional, but clearly it can't go
anywhere for a while yet regardless. Will seek clarification. No review.