Reworking TypeMaps to avoid allocations.
The pieces of TypeMap which returned Option[T] were contributing many
unnecessary allocations as well as inconsistency and complexity. By
creating distinguished objects UnmappableTree and UnmappableAnnotation,
I was able to gain simplicity, eliminate all the Some allocations, use
the normal mapConserve, and so on.

I also went after the monster that was dropNonConStraintAnnotations by
generalizing it into "AnnotationFilter", a trait which extends TypeMap.
SpecializeTypes was able to take advantage of this as well. (The means
by which (some) annotations are dropped, and the related fact that
NullaryMethodType returns constant false for isTrivial only so the
continuations plugin will keep working, leaves the situation still in
dire need of improvement.)

Review by moors.