Fixing isTrivial in NullaryMethod.
1) Made NullaryMethodType a SimpleTypeProxy. (This is incidental to
isTrivial, it just seemed like the right thing to do.)

2) Modified implementations of isTrivial for MethodType and
NullaryMethodType to only be considered trivial if there are no
annotations on the return type. This means asSeenFrom will not be
skipped and the continuations plugin will work properly. There are a lot
of other angles one could take, so review by moors, rompf.