Reflection classloader adjustment.
With this change, one can use the reflective compiler with the scala
libs on the boot classpath (the default) or not (with -nobootcp) and it
will work either way. Maybe it will work in sbt now too, but I haven't
tried it. Since I bet you will try it, review by grek.