Fixing checkinit build.
As near as I can tell, -Xcheckinit was failing because a var was read
too early as null rather than waiting for it to be set to its initial
value, null. I guess that's a defensible warning, since one might change
it away from null. Organized the vars in that file better. No review.