Made LazyType fail more graciously with incompl...
Made LazyType fail more graciously with incomplete class paths.
SymbolLoaders already do the right thing, printing an error instead of
crashing the compiler. However, the unpickler has two more lazy types
that crash and stop the compiler if, on completion, it encounters a
reference to a missing class file.

Since the SymbolTable has no way of reporting an error, we convert the
MissingRequirementError in TypeErrors. This has the benefit of being
printed nicely by the type checker (with a position where the type was
needed).

This fixes extremely worrying behavior in the IDE when the class path is
incomplete (for instance, after a rebuild with compilation errors).

review by odersky, extempore.