Enable the use of spaces in paths for the Scala...
Enable the use of spaces in paths for the Scala build on Windows.

Revert r25995 which was fixing it only partly and in the wrong place.
Properly encode argument files for scalac in scalac ant task. Allow
'compilerarg' elements in scalac ant task (like in ant's built-in javac
task) to allow passing extra parameters like plugindir path with proper
encoding of spaces and file names, and use it in the Scala build. Fix
space handling in get-scala-revision.bat.

(Patch by Stefan Zeiger.) Closes SI-3047.