Enable the use of spaces in paths for the Scala...
Enable the use of spaces in paths for the Scala build on Windows -- take 2.

(The original commit in r26026, reverted in r26027, used the new
compilerargs

element in the Scala build -- we cannot do this until it's in starr.)

- Revert r25995 which was fixing it only partly and in the wrong place.
- Properly encode argument files for scalac in scalac ant task. - Allow
'compilerarg' elements in scalac ant task (like in ant's built-in
javac task) to allow passing extra parameters like plugindir path with
proper encoding of spaces and file names. - Fix space handling in
get-scala-revision.bat.

Closes SI-3047.