Re-enable RedBlack's ScalaCheck test, which was disabled at r21829.
I haven't been able to find any reason for that reversal, but
this version is an improved version provided with SI-3796, which
never got committed along with the patch.

This version generates trees directly instead of using the methods
under test. It also puts an upper bound on tree size and, therefore,
execution time.