Use git-describe to create build string.
Created dev tag which is the merge-base of master and 2.9.1 (merge-base
is the last common commit), and told ant to generate build strings based
on that. So distributions look like

  scala-2.10.0.dev-1289-gbba3918

Where the 1289 means you are 1289 commits ahead of "dev" and
the last segment (minus the g) is the sha-1 hash.

This no doubt breaks windows even further.  Help getting it
back on its feet greatly appreciated.