Tweaks to get-scala-revision.
Reference to git head is supposed to be HEAD, not head.
Also apparently [[ <test> ]] is a bashism so it's [ <test> ].
Made name generation robust against the absence of the dev
tag; it will generate a similar name even if it's not there,
minus the part it can't determine without it.