Changed shadowed constructor parameter names.
One of my more recent warnings tells us this a bunch of places
if we compile with -Xlint, one example:

  scala/actors/ReplyReactorTask.scala:26: warning:
    private[this] value reactor in class ReplyReactorTask shadows
    mutable reactor inherited from class ReactorTask. Changes to reactor
    will not be visible within class ReplyReactorTask - you may want to
    give them distinct names.

In some cases (like that one) I was not confident whether the shadowing
was intentional, in which case I left the semantics but changed the name
anyway because it will be eternally confusing otherwise.

Review by @phaller.