Avoid calling apply twice on Range.foreach.
On micro-benchmarks, this resulted in slightly slower foreach loops
for small ranges (10-sized loops), but some gains for bigger loops
(100-sized+). Small loops with Range are extremely slow compared to
while loops anyway.

More importantly, it reduces the size of inlined code with -optimise,
which increases the chance of inlining happening and of JVM further
improving upon it.

	Fixes SI-5286