typeParams now performs one more completion
When tracking a problem in type inference during reflective compilation,
Martin and I have discovered that under certain circumstances typeParams
does not fully load the type.

During a debugging session we found out that info first had the type of
TopLevelCompleter, and after the load it was still not completed, having
the type of UnPickler$something$LazyTypeRef.

After that discovery it became apparent that typeParams need to behave
similarly to the info getter - allow for two completions, not just one.

Review by @odersky, @adriaanm.