Fix bugs with lubs.
Was not substituting before using fallback lub.  I don't think
it was hurting anything because the substitution is wrong anyway,
since it gives up on f-bounds.  (This is directly responsible
for the malformed lubs.) But if it must fail, it shouldn't be
letting A and Repr escape into the error message:

scala> List(List(), Stream())
<console>:8: error: type mismatch;
 found   : List[Nothing]
 required: scala.collection.immutable.LinearSeq[Nothing] with .....
   scala.collection.LinearSeqOptimized[A,Repr]]]]]
             List(List(), Stream())
                      ^
Rather, like this:

 found   : List[Nothing]
 required: scala.collection.immutable.LinearSeq[Nothing] with .....
   scala.collection.LinearSeqOptimized[Nothing,scala.collection.immutable.Stream[Nothing]]]]]]

Also, AbstractSeq and other not-public classes were appeaing in the
errors, alerting me to the fact that it would happily infer a lub
containing a type which cannot be referenced.

  scala> List(List(), Vector())
  res1: List[scala.collection.immutable.Seq[Nothing] with scala.collection.AbstractSeq[Nothing] ...

So I excluded non-public types in lubList:

  scala> List(List(), Vector())
  res0: List[scala.collection.immutable.Seq[Nothing]] = List(List(), Vector())

Finally, I added logging for recursive bounds.

% scala -Ydebug
scala> List(List(), Stream())
Encountered 2 recursive bounds while lubbing 2 types.
  type A in trait LinearSeqOptimized appears in the bounds of type Repr
    type Repr >: scala.this.Nothing <: collection.this.LinearSeqOptimized[A,Repr]
  type Repr in trait LinearSeqOptimized appears in its own bounds
    type Repr >: scala.this.Nothing <: collection.this.LinearSeqOptimized[A,Repr]

Review by @moors.