Made type constraint propagation off by default.
Wow, it fails only during strap.comp. That is certainly a first for me,
managing to build quick and pass all the tests only to be unable to
rebuild the compiler from quick. Must be something fascinating taking
place in there.  So for now I flipped the switch back down.