Simplified emitSwitch in PatMatVirtualiser.
Simplified emitSwitch method that does not emit
forward jumps for switch tables. Instead, it
emits Match trees that are of a shape that
can be directly translated to switch.

The backend takes care of the actual translation
from Match trees into switch tables.

As a bonus, we emit more compact trees compared
to the old implementation of emitSwitch.

Review by @adriaanm.