Scalac fork no longer dumps stacktraces on compilation errors
Current behavior of scalacfork task is to fail the build when there are
compilation errors reported by scalac fork. So far, so good.

However, this functionality is implemented by throwing sys.error, which
makes ant dump the entire stacktrace. This is annoying, since it almost
certainly scrolls the screen away of the error (hello, dear 1366x768)
and buries it under a meaningless stacktrace.

Surprisingly, there is a very simple fix that remedies the situation.
Credit goes to @bakoyaro from SO: http://bit.ly/xdR306